Revert "Revert "Merge pull request #582 from pmeenan/safari""
This reverts commit ea286035c1a8522fc23ad655270a5614a6529f71.
