--- internal/support/Safari/extension/wptagent/wptagent Extension/SafariWebExtensionHandler.swift
@@ -1,26 +1,0 @@
-//
-//  SafariWebExtensionHandler.swift
-//  wptagent Extension
-//
-//  Created by Wuyang Wang on 9/28/20.
-//
-
-import SafariServices
-import os.log
-
-let SFExtensionMessageKey = "message"
-
-class SafariWebExtensionHandler: NSObject, NSExtensionRequestHandling {
-
-	func beginRequest(with context: NSExtensionContext) {
-        let item = context.inputItems[0] as! NSExtensionItem
-        let message = item.userInfo?[SFExtensionMessageKey]
-        os_log(.default, "Received message from browser.runtime.sendNativeMessage: %@", message as! CVarArg)
-
-        let response = NSExtensionItem()
-        response.userInfo = [ SFExtensionMessageKey: [ "Response to": message ] ]
-
-        context.completeRequest(returningItems: [response], completionHandler: nil)
-    }
-    
-}
