--- internal/support/Safari/extension/wptagent/wptagent/AppDelegate.swift
@@ -1,21 +1,0 @@
-//
-//  AppDelegate.swift
-//  wptagent
-//
-//  Created by Wuyang Wang on 9/28/20.
-//
-
-import Cocoa
-
-@NSApplicationMain
-class AppDelegate: NSObject, NSApplicationDelegate {
-    
-    func applicationDidFinishLaunching(_ aNotification: Notification) {
-        // Insert code here to initialize your application
-    }
-    
-    func applicationWillTerminate(_ aNotification: Notification) {
-        // Insert code here to tear down your application
-    }
-    
-}
